# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/git/hook": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Git"
                ],
                "summary": "Git钩子",
                "parameters": [
                    {
                        "description": "Git钩子",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.HookReq"
                        }
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "400": {
                        "description": "错误的请求",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "500": {
                        "description": "系统异常",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "requests.HookRepository": {
            "type": "object",
            "properties": {
                "git_http_url": {
                    "type": "integer"
                },
                "git_ssh_url": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "requests.HookReq": {
            "type": "object",
            "properties": {
                "event_name": {
                    "type": "string"
                },
                "ref": {
                    "type": "integer"
                },
                "repository": {
                    "$ref": "#/definitions/requests.HookRepository"
                },
                "user_email": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                },
                "user_name": {
                    "type": "integer"
                }
            }
        },
...
        "responses.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}